Updated usage, changed c_size_t to int40, updated test_queries()

+++ Recompression/typedefs.hpp
@@ -2,8 +2,9 @@
 #define TYPEDEFS_HPP
 
 #include <cstdint>
+#include "types/int40.hpp"
 
-typedef int32_t c_size_t;
+typedef int40 c_size_t;
 typedef uint32_t uc_size_t;
 typedef int64_t len_t;
 typedef uint8_t char_t;
